+++ src/store/Provider.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Provider } from "react-redux";
+import store from "./store";
+
+interface ProviderProps {
+    
+}
+
+export const ReduxProvider: React.FC = ({  }: ProviderProps) => {
+    return <Provider store={store}>
+
+    </Provider>
+};
+
+export default ReduxProvider;